Moved Obstacles to another class
